Add make_finalised_mt job
