+++ joint_calling/variant_qc.py
@@ -1,0 +1,323 @@
+"""
+Variant QC related hail-query jobs
+"""
+
+import uuid
+from os.path import join
+from typing import List, Optional, Dict, Tuple
+import logging
+import pandas as pd
+import hailtop.batch as hb
+from hailtop.batch.job import Job
+from analysis_runner import dataproc
+
+from joint_calling import utils
+from joint_calling.vqsr import make_vqsr_jobs
+
+logger = logging.getLogger('joint-calling')
+logger.setLevel('INFO')
+
+
+def add_variant_qc_jobs(
+    b: hb.Batch,
+    work_bucket: str,
+    analysis_bucket: str,
+    raw_combined_mt_path: str,
+    info_split_ht_path: str,
+    hard_filter_ht_path: str,
+    meta_ht_path: str,
+    samples_df: pd.DataFrame,
+    sample_qc_job: Job,
+    combiner_job: Job,
+    scripts_dir: str,
+    ped_file: Optional[str],
+    overwrite: bool,
+    run_rf: bool,
+    vqsr_params_d: Dict,
+) -> Tuple[Job, str]:
+    """
+    Add variant QC related hail-query jobs
+    """
+    rf_bucket = join(work_bucket, 'rf')
+    vqsr_bucket = join(work_bucket, 'vqsr')
+
+    fam_stats_ht_path = join(work_bucket, 'fam-stats.ht')
+    allele_data_ht_path = join(work_bucket, 'allele-data.ht')
+    qc_ac_ht_path = join(work_bucket, 'qc-ac.ht')
+    rf_annotations_ht_path = None
+
+    if overwrite or not all(
+        utils.file_exists(fp) for fp in [allele_data_ht_path, qc_ac_ht_path]
+    ):
+        anno_job = dataproc.hail_dataproc_job(
+            b,
+            f'{scripts_dir}/generate_rf_annotations.py --overwrite '
+            + f'--mt {raw_combined_mt_path} '
+            + f'--hard-filtered-samples-ht {hard_filter_ht_path} '
+            + f'--meta-ht {meta_ht_path} '
+            + f'--out-allele-data-ht {allele_data_ht_path} '
+            + f'--out-qc-ac-ht {qc_ac_ht_path} '
+            + f'--out-fam-stats-ht {fam_stats_ht_path} '
+            + (f'--fam-file {ped_file} ' if ped_file else '')
+            + f'--bucket {work_bucket} ',
+            max_age='8h',
+            packages=utils.DATAPROC_PACKAGES,
+            num_secondary_workers=10,
+            depends_on=[sample_qc_job],
+            job_name='Var QC: generate QC annotations',
+            vep='GRCh38',
+        )
+    else:
+        anno_job = b.new_job('Var QC: generate QC annotations [reuse]')
+
+    freq_ht_path = join(work_bucket, 'frequencies.ht')
+    if overwrite or not utils.file_exists(freq_ht_path):
+        freq_job = dataproc.hail_dataproc_job(
+            b,
+            f'{scripts_dir}/generate_freq_data.py --overwrite '
+            f'--mt {raw_combined_mt_path} '
+            f'--hard-filtered-samples-ht {hard_filter_ht_path} '
+            f'--meta-ht {meta_ht_path} '
+            f'--out-ht {freq_ht_path} '
+            f'--bucket {work_bucket} ',
+            max_age='8h',
+            packages=utils.DATAPROC_PACKAGES,
+            num_secondary_workers=10,
+            depends_on=[sample_qc_job],
+            job_name='Var QC: generate frequencies',
+        )
+    else:
+        freq_job = b.new_job('RF: generate frequencies [reuse]')
+
+    if run_rf:
+        rf_annotations_ht_path = join(work_bucket, 'rf-annotations.ht')
+        rf_result_ht_path = join(work_bucket, 'rf-result.ht')
+        rf_model_id = f'rf_{str(uuid.uuid4())[:8]}'
+        if overwrite or not any(
+            utils.file_exists(path)
+            for path in [rf_annotations_ht_path, rf_result_ht_path]
+        ):
+            rf_job = dataproc.hail_dataproc_job(
+                b,
+                f'{scripts_dir}/random_forest.py --overwrite '
+                f'--info-split-ht {info_split_ht_path} '
+                f'--freq-ht {freq_ht_path} '
+                f'--allele-data-ht {allele_data_ht_path} '
+                f'--qc-ac-ht {qc_ac_ht_path} '
+                f'--bucket {work_bucket} '
+                f'--use-adj-genotypes '
+                f'--out-model-id {rf_model_id} '
+                f'--out-annotations-ht {rf_annotations_ht_path} '
+                f'--out-ht {rf_result_ht_path} ',
+                max_age='8h',
+                packages=utils.DATAPROC_PACKAGES,
+                num_secondary_workers=10,
+                depends_on=[freq_job, anno_job],
+                job_name='Random forest ',
+            )
+        else:
+            rf_job = b.new_job('Random forest [reuse]')
+
+        final_job, final_filter_ht_path = make_rf_eval_jobs(
+            b=b,
+            combined_mt_path=raw_combined_mt_path,
+            info_split_ht_path=info_split_ht_path,
+            rf_result_ht_path=rf_result_ht_path,
+            rf_annotations_ht_path=rf_annotations_ht_path,
+            fam_stats_ht_path=fam_stats_ht_path,
+            freq_ht_path=freq_ht_path,
+            rf_model_id=rf_model_id,
+            work_bucket=rf_bucket,
+            overwrite=overwrite,
+            scripts_dir=scripts_dir,
+            depends_on=[rf_job],
+        )
+
+    else:
+        final_gathered_vcf_job, final_gathered_vcf_path = make_vqsr_jobs(
+            b,
+            combined_mt_path=raw_combined_mt_path,
+            hard_filter_ht_path=hard_filter_ht_path,
+            meta_ht_path=meta_ht_path,
+            gvcf_count=len(samples_df),
+            vqsr_bucket=vqsr_bucket,
+            analysis_bucket=join(analysis_bucket, 'vqsr'),
+            depends_on=[combiner_job],
+            scripts_dir=scripts_dir,
+            vqsr_params_d=vqsr_params_d,
+        )
+        final_job, final_filter_ht_path = make_vqsr_eval_jobs(
+            b=b,
+            combined_mt_path=raw_combined_mt_path,
+            info_split_ht_path=info_split_ht_path,
+            final_gathered_vcf_path=final_gathered_vcf_path,
+            rf_annotations_ht_path=rf_annotations_ht_path,
+            fam_stats_ht_path=fam_stats_ht_path,
+            freq_ht_path=freq_ht_path,
+            work_bucket=vqsr_bucket,
+            analysis_bucket=join(analysis_bucket, 'vqsr'),
+            overwrite=overwrite,
+            scripts_dir=scripts_dir,
+            depends_on=[anno_job, freq_job, final_gathered_vcf_job],
+        )
+    return final_job, final_filter_ht_path
+
+
+def make_rf_eval_jobs(
+    b: hb.Batch,
+    combined_mt_path: str,
+    info_split_ht_path: str,
+    rf_result_ht_path: str,
+    rf_annotations_ht_path: str,
+    fam_stats_ht_path: str,
+    freq_ht_path: str,
+    rf_model_id: str,
+    work_bucket: str,
+    overwrite: bool,
+    scripts_dir: str,
+    depends_on: Optional[List[Job]],
+) -> Tuple[Job, str]:
+    """
+    Make jobs that do evaluation RF model and applies the final filters
+
+    Returns the final_filter Job object and the path to the final filter HT
+    """
+    score_bin_ht_path = join(work_bucket, 'rf-score-bin.ht')
+    score_bin_agg_ht_path = join(work_bucket, 'rf-score-agg-bin.ht')
+    if overwrite or not utils.file_exists(score_bin_ht_path):
+        eval_job = dataproc.hail_dataproc_job(
+            b,
+            f'{scripts_dir}/evaluation.py --overwrite '
+            f'--info-split-ht {info_split_ht_path} '
+            f'--rf-results-ht {rf_result_ht_path} '
+            f'--rf-annotations-ht {rf_annotations_ht_path} '
+            f'--fam-stats-ht {fam_stats_ht_path} '
+            f'--mt {combined_mt_path} '
+            f'--bucket {work_bucket} '
+            f'--out-bin-ht {score_bin_ht_path} '
+            f'--out-aggregated-bin-ht {score_bin_agg_ht_path} '
+            f'--run-sanity-checks ',
+            max_age='8h',
+            packages=utils.DATAPROC_PACKAGES,
+            num_secondary_workers=10,
+            depends_on=depends_on,
+            job_name='RF: evaluation',
+        )
+    else:
+        eval_job = b.new_job('RF: evaluation [reuse]')
+
+    final_filter_ht_path = join(work_bucket, 'final-filter.ht')
+    if not utils.file_exists(final_filter_ht_path):
+        final_filter_job = dataproc.hail_dataproc_job(
+            b,
+            f'{scripts_dir}/final_filter.py --overwrite '
+            f'--out-final-filter-ht {final_filter_ht_path} '
+            f'--model-id {rf_model_id} '
+            f'--model-name RF '
+            f'--score-name RF '
+            f'--info-split-ht {info_split_ht_path} '
+            f'--freq-ht {freq_ht_path} '
+            f'--score-bin-ht {score_bin_ht_path} '
+            f'--score-bin-agg-ht {score_bin_agg_ht_path} ' + f'--bucket {work_bucket} ',
+            max_age='8h',
+            packages=utils.DATAPROC_PACKAGES,
+            num_secondary_workers=10,
+            depends_on=[eval_job],
+            job_name='RF: final filter',
+        )
+    else:
+        final_filter_job = b.new_job('RF: final filter [reuse]')
+
+    return final_filter_job, final_filter_ht_path
+
+
+def make_vqsr_eval_jobs(
+    b: hb.Batch,
+    combined_mt_path: str,
+    info_split_ht_path: str,
+    final_gathered_vcf_path: str,
+    rf_annotations_ht_path: Optional[str],
+    fam_stats_ht_path: str,
+    freq_ht_path: str,
+    work_bucket: str,
+    analysis_bucket: str,  # pylint: disable=unused-argument
+    overwrite: bool,
+    scripts_dir: str,
+    depends_on: Optional[List[Job]],
+) -> Tuple[Job, str]:
+    """
+    Make jobs that do evaluation VQSR model and applies the final filters
+
+    Returns the final_filter Job object and the path to the final filter HT
+    """
+    vqsr_filters_split_ht_path = join(work_bucket, 'vqsr-filters-split.ht')
+    if overwrite or not utils.file_exists(vqsr_filters_split_ht_path):
+        load_vqsr_job = dataproc.hail_dataproc_job(
+            b,
+            f'{scripts_dir}/load_vqsr.py --overwrite '
+            f'--split-multiallelic '
+            f'--out-path {vqsr_filters_split_ht_path} '
+            f'--vqsr-vcf-path {final_gathered_vcf_path} '
+            f'--bucket {work_bucket} ',
+            max_age='8h',
+            packages=utils.DATAPROC_PACKAGES,
+            num_secondary_workers=10,
+            depends_on=depends_on,
+            job_name='VQSR: load_vqsr',
+        )
+    else:
+        load_vqsr_job = b.new_job('VQSR: load_vqsr')
+
+    score_bin_ht_path = join(work_bucket, 'vqsr-score-bin.ht')
+    score_bin_agg_ht_path = join(work_bucket, 'vqsr-score-agg-bin.ht')
+    if overwrite or not utils.file_exists(score_bin_ht_path):
+        eval_job = dataproc.hail_dataproc_job(
+            b,
+            f'{scripts_dir}/evaluation.py --overwrite '
+            f'--info-split-ht {info_split_ht_path} '
+            + (
+                f'--rf-annotations-ht {rf_annotations_ht_path} '
+                if rf_annotations_ht_path
+                else ''
+            )
+            + f'--fam-stats-ht {fam_stats_ht_path} '
+            f'--vqsr-filters-split-ht {vqsr_filters_split_ht_path} '
+            f'--mt {combined_mt_path} '
+            f'--bucket {work_bucket} '
+            f'--out-bin-ht {score_bin_ht_path} '
+            f'--out-aggregated-bin-ht {score_bin_agg_ht_path} '
+            f'--run-sanity-checks ',
+            max_age='8h',
+            packages=utils.DATAPROC_PACKAGES,
+            num_secondary_workers=10,
+            depends_on=[load_vqsr_job],
+            job_name='VQSR: evaluation',
+        )
+    else:
+        eval_job = b.new_job('VQSR: evaluation')
+
+    final_filter_ht_path = join(work_bucket, 'final-filter.ht')
+    vqsr_model_id = 'vqsr_model'
+    if not utils.file_exists(final_filter_ht_path):
+        final_filter_job = dataproc.hail_dataproc_job(
+            b,
+            f'{scripts_dir}/final_filter.py --overwrite '
+            f'--out-final-filter-ht {final_filter_ht_path} '
+            f'--vqsr-filters-split-ht {vqsr_filters_split_ht_path} '
+            f'--model-id {vqsr_model_id} '
+            f'--model-name VQSR '
+            f'--score-name AS_VQSLOD '
+            f'--info-split-ht {info_split_ht_path} '
+            f'--freq-ht {freq_ht_path} '
+            f'--score-bin-ht {score_bin_ht_path} '
+            f'--score-bin-agg-ht {score_bin_agg_ht_path} '
+            f'--bucket {work_bucket} ',
+            max_age='8h',
+            packages=utils.DATAPROC_PACKAGES,
+            num_secondary_workers=10,
+            depends_on=[eval_job],
+            job_name='VQSR: final filter',
+        )
+    else:
+        final_filter_job = b.new_job('VQSR: final filter')
+    return final_filter_job, final_filter_ht_path
