+++ joint_calling/vqsr.py
@@ -363,92 +363,6 @@ def make_vqsr_jobs(
     return final_gathered_vcf_job, final_gathered_vcf['vcf.gz']
 
 
-def _make_vqsr_eval_jobs(
-    b: hb.Batch,
-    combined_mt_path: str,
-    info_split_ht_path: str,
-    final_gathered_vcf_path: str,
-    rf_annotations_ht_path: str,
-    fam_stats_ht_path: str,
-    freq_ht_path: str,
-    work_bucket: str,
-    analysis_bucket: str,  # pylint: disable=unused-argument
-    overwrite: bool,
-    scripts_dir: str,
-    depends_on: Optional[List[Job]],
-) -> Job:
-    """
-    Make jobs that do evaluation VQSR model and applies the final filters
-    """
-    vqsr_filters_split_ht_path = join(work_bucket, 'vqsr-filters-split.ht')
-    if overwrite or not utils.file_exists(vqsr_filters_split_ht_path):
-        load_vqsr_job = dataproc.hail_dataproc_job(
-            b,
-            f'{scripts_dir}/load_vqsr.py --overwrite '
-            f'--split-multiallelic '
-            f'--out-path {vqsr_filters_split_ht_path} '
-            f'--vqsr-vcf-path {final_gathered_vcf_path} '
-            f'--bucket {work_bucket} ',
-            max_age='8h',
-            packages=utils.DATAPROC_PACKAGES,
-            num_secondary_workers=10,
-            depends_on=depends_on,
-            job_name='VQSR: load_vqsr',
-        )
-    else:
-        load_vqsr_job = b.new_job('VQSR: load_vqsr')
-
-    score_bin_ht_path = join(work_bucket, 'vqsr-score-bin.ht')
-    score_bin_agg_ht_path = join(work_bucket, 'vqsr-score-agg-bin.ht')
-    if overwrite or not utils.file_exists(score_bin_ht_path):
-        eval_job = dataproc.hail_dataproc_job(
-            b,
-            f'{scripts_dir}/evaluation.py --overwrite '
-            f'--info-split-ht {info_split_ht_path} '
-            f'--rf-annotations-ht {rf_annotations_ht_path} '
-            f'--fam-stats-ht {fam_stats_ht_path} '
-            f'--vqsr-filters-split-ht {vqsr_filters_split_ht_path} '
-            f'--mt {combined_mt_path} '
-            f'--bucket {work_bucket} '
-            f'--out-bin-ht {score_bin_ht_path} '
-            f'--out-aggregated-bin-ht {score_bin_agg_ht_path} '
-            f'--run-sanity-checks ',
-            max_age='8h',
-            packages=utils.DATAPROC_PACKAGES,
-            num_secondary_workers=10,
-            depends_on=[load_vqsr_job],
-            job_name='VQSR: evaluation',
-        )
-    else:
-        eval_job = b.new_job('VQSR: evaluation')
-
-    final_filter_ht_path = join(work_bucket, 'final-filter.ht')
-    vqsr_model_id = 'vqsr_model'
-    if not utils.file_exists(final_filter_ht_path):
-        final_filter_job = dataproc.hail_dataproc_job(
-            b,
-            f'{scripts_dir}/final_filter.py --overwrite '
-            f'--out-final-filter-ht {final_filter_ht_path} '
-            f'--vqsr-filters-split-ht {vqsr_filters_split_ht_path} '
-            f'--model-id {vqsr_model_id} '
-            f'--model-name VQSR '
-            f'--score-name AS_VQSLOD '
-            f'--info-split-ht {info_split_ht_path} '
-            f'--freq-ht {freq_ht_path} '
-            f'--score-bin-ht {score_bin_ht_path} '
-            f'--score-bin-agg-ht {score_bin_agg_ht_path} '
-            f'--bucket {work_bucket} ',
-            max_age='8h',
-            packages=utils.DATAPROC_PACKAGES,
-            num_secondary_workers=10,
-            depends_on=[eval_job],
-            job_name='VQSR: final filter',
-        )
-    else:
-        final_filter_job = b.new_job('VQSR: final filter')
-    return final_filter_job
-
-
 def add_tabix_step(
     b: hb.Batch,
     vcf_path: str,
