+++ scripts/evaluation.py
@@ -57,9 +57,8 @@ logger.setLevel('INFO')
 @click.option(
     '--rf-annotations-ht',
     'rf_annotations_ht_path',
-    required=True,
     callback=get_validation_callback(ext='ht'),
-    help='RF annotations table (created by generate_rf_annotations.py)',
+    help='RF annotations table (created by random_forest.py)',
 )
 @click.option(
     '--mt',
@@ -131,11 +130,11 @@ logger.setLevel('INFO')
 )
 def main(  # pylint: disable=too-many-arguments,too-many-locals
     info_split_ht_path: str,
-    rf_annotations_ht_path: str,
     mt_path: str,
     fam_stats_ht_path: str,
     rf_results_ht_path: Optional[str],
     vqsr_filters_split_ht_path: Optional[str],
+    rf_annotations_ht_path: Optional[str],
     out_bin_ht: str,
     out_aggregated_bin_ht: str,
     run_sanity_checks: bool,
@@ -153,13 +152,15 @@ def main(  # pylint: disable=too-many-arguments,too-many-locals
     if overwrite or not utils.file_exists(out_bin_ht):
         scores_ht = create_bin_ht(
             info_split_ht=hl.read_table(info_split_ht_path),
-            rf_annotations_ht=hl.read_table(rf_annotations_ht_path),
             n_bins=n_bins,
             rf_results_ht=hl.read_table(rf_results_ht_path)
             if rf_results_ht_path
             else None,
             vqsr_filters_split_ht=hl.read_table(vqsr_filters_split_ht_path)
             if vqsr_filters_split_ht_path
+            else None,
+            rf_annotations_ht=hl.read_table(rf_annotations_ht_path)
+            if rf_annotations_ht_path
             else None,
         )
         scores_ht.write(out_bin_ht, overwrite=True)
@@ -354,30 +355,32 @@ def _truth_concordance(
 
 def create_bin_ht(
     info_split_ht: hl.Table,
-    rf_annotations_ht: hl.Table,
     n_bins: int,
     rf_results_ht: Optional[hl.Table] = None,
     vqsr_filters_split_ht: Optional[hl.Table] = None,
+    rf_annotations_ht: Optional[hl.Table] = None,
 ) -> hl.Table:
     """
     Creates a table with bin annotations added for a RF or VQSR run and writes it
     to its correct location in annotations.
 
     :param info_split_ht: table generated by generate_info_ht.py
-    :param rf_results_ht: table generated by random_forest.py
-    :param rf_annotations_ht: table generated by random_forest.py
-    :param vqsr_filters_split_ht: table generated by load_vqsr.py
     :param n_bins: Number of bins to bin the data into
+    :param rf_results_ht: table generated by random_forest.py
+    :param vqsr_filters_split_ht: table generated by load_vqsr.py
+    :param rf_annotations_ht: table generated by random_forest.py
     :return: Table with bin annotations
     """
     logger.info(f'Annotating model HT with bins using {n_bins} bins')
     if vqsr_filters_split_ht:
         logger.info(f'Using a VQSR model')
-        rf_ht = rf_annotations_ht
+
         ht = vqsr_filters_split_ht
 
+        if rf_annotations_ht is not None:
+            ht = ht.annotate(**rf_annotations_ht[ht.key])
+
         ht = ht.annotate(
-            **rf_ht[ht.key],
             info=info_split_ht[ht.key].info,
             score=ht.info.AS_VQSLOD,
             positive_train_site=ht.info.POSITIVE_TRAIN_SITE,
@@ -385,7 +388,7 @@ def create_bin_ht(
             AS_culprit=ht.info.AS_culprit,
         )
 
-        # Remove all samples with an undefined ac_raw Because ac_raw was calculated
+        # Remove all samples with an undefined ac_raw because ac_raw was calculated
         # on the high quality samples only and VQSR was run before sample filtering
         ht = ht.filter(hl.is_defined(ht.ac_raw))
